Own Projects/api/requirements.txt
cached-property==1.5.2
click==8.1.3
emoji==1.7.0
Flask==2.1.2
Flask-Cors==3.0.10
grpcio==1.46.1
gunicorn==20.1.0
h5py==3.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.21.6
pickle5==0.0.11
pip==22.1
regex==2022.4.24
setuptools==62.2.0
six==1.16.0
tqdm==4.64.0
tensorflow-cpu==2.8.0
Werkzeug==2.1.2
wheel==0.37.1
